requirements.txt
cycler==0.10.0
drawnow==0.72.0
kiwisolver==1.0.1
matplotlib==3.0.3
pyparsing==2.4.0
pyserial==3.4
python-dateutil==2.8.0
RFExplorer==1.26.1806.1
six==1.12.0
